/feature_route_details_editor PrivateRoutesFragment.kt fix getting annotation data from asInt to asLong

+++ app/src/main/java/com/example/gh_coursework/ui/private_route/PrivateRoutesFragment.kt
@@ -125,7 +125,7 @@ class PrivateRoutesFragment :
         if (annotation.getData()?.isJsonNull == false) {
 
             viewLifecycleOwner.lifecycleScope.launch {
-                annotation.getData()?.asInt?.let { pointId ->
+                annotation.getData()?.asLong?.let { pointId ->
                     viewModel.getPointDetailsPreview(pointId).collect { details ->
                         prepareDetailsDialog(annotation, details)
                     }
@@ -775,7 +775,7 @@ class PrivateRoutesFragment :
 
     private fun loadPointData(annotation: PointAnnotation) {
         viewLifecycleOwner.lifecycleScope.launch {
-            annotation.getData()?.asInt?.let { pointId ->
+            annotation.getData()?.asLong?.let { pointId ->
                 viewModel.getPointDetailsPreview(pointId).collect { details ->
                     prepareDetailsDialog(annotation, details)
                 }
